take a look
Hi Satvik, I have been working these week on this script, it takes the data from a csv file df_oc.csv and moves the qubit with the predictions from the model.
There is one change you have to do and is to change the path for the csv in the async_bloch.py file at the end.

The script is an offline version of what we should try to do.
Once you can look at it and see where to connect your headset to the model you can try and see the results online.
If you need any comment on the code pls ask me. I will try to answer as fast as possible. 
See you on monday.