pls check my comments & email
Hi Satvik, as I already mentioned in the mail I have several comments on the code.

The shape of the feature vector should be (1, number_of_features) being number_of_features the number of features used to train the model.

Also the model that I currently have is the open/closed KNeighborsClassifier so I'm not sure that it will work in this case but, as long as they train their own model following our instructions it should be ok.

Best,
Carlos.